+++ JAPAN_GRANDPRIZE_2020/Map.h
@@ -3,4 +3,4 @@
 void MapDisp();	// マップの表示
 void MapMove();	// マップの動き
 void MapInit();	// マップの初期化
-void Scroll(float* mapX);	// スクロール処理
+void Scroll(int* mapX);	// スクロール処理
